+++ packages/node/src/behavior/state/managed/Datasource.ts
@@ -26,7 +26,7 @@ const logger = Logger.get("Datasource");
 
 const FEATURES_KEY = "__features__";
 
-const stateChanged = Symbol("stateChanged");
+const changed = Symbol("changed");
 
 const viewTx = Transaction.open("offline-view", "ro");
 
@@ -66,7 +66,7 @@ export interface Datasource<T extends StateType = StateType> extends Transaction
     /**
      * Event that gets emitted when the state changes.
      */
-    stateChanged: Observable<[changes: string[], version: number], MaybePromise>;
+    changed: Observable<[changes: string[], version: number], MaybePromise>;
 
     /**
      * Events registered for this Datasource
@@ -103,8 +103,8 @@ export function Datasource<const T extends StateType = StateType>(options: Datas
             return internals.location;
         },
 
-        get stateChanged() {
-            return internals.events[stateChanged];
+        get changed() {
+            return internals.events[changed];
         },
 
         get events() {
@@ -144,13 +144,13 @@ export namespace Datasource {
     export type Events = {
         interactionBegin?: Observable<[context?: ValueSupervisor.Session], MaybePromise>;
         interactionEnd?: Observable<[context?: ValueSupervisor.Session], MaybePromise>;
-        datasourceChanged?: Observable<[context?: ValueSupervisor.Session], MaybePromise>;
+        stateChanged?: Observable<[context?: ValueSupervisor.Session], MaybePromise>;
     } & {
         [K in `${string}$Changing` | `${string}$Changed`]: Observable<Parameters<ValueObserver>, MaybePromise>;
     };
 
     export type InternalEvents = Events & {
-        [stateChanged]: Observable<[changes: string[], version: number], MaybePromise>;
+        [changed]: Observable<[changes: string[], version: number], MaybePromise>;
     };
 
     /**
@@ -286,7 +286,7 @@ function configure(options: Datasource.Options): Internals {
     Object.freeze(options.location);
 
     const events = (options.events ?? {}) as Datasource.InternalEvents;
-    events[stateChanged] = new Observable();
+    events[changed] = new Observable();
 
     return {
         ...options,
@@ -624,9 +624,9 @@ function createReference(resource: Transaction.Resource, internals: Internals, s
             // We don't revert the version number on rollback.  Should be OK
             incrementVersion();
 
-            if (internals.events.datasourceChanged?.isObserved) {
+            if (internals.events.stateChanged?.isObserved) {
                 changes.notifications.push({
-                    event: internals.events.datasourceChanged,
+                    event: internals.events.stateChanged,
                     params: [session],
                 });
             }
@@ -695,7 +695,7 @@ function createReference(resource: Transaction.Resource, internals: Internals, s
             }
         }
 
-        const changeSetResult = internals.events[stateChanged]?.emit(
+        const changeSetResult = internals.events[changed]?.emit(
             Array.from(changes.changeList.values()),
             internals.version,
         );