+++ packages/protocol/src/fabric/Fabric.ts
@@ -26,8 +26,8 @@ import {
     PrivateKey,
     StorageContext,
 } from "#general";
-import { FabricGroupsManager, GROUP_SECURITY_INFO } from "#groups/FabricGroupsManager.js";
-import { FabricAccessControlManager } from "#interaction/FabricAccessControlManager.js";
+import { FabricGroups, GROUP_SECURITY_INFO } from "#groups/FabricGroups.js";
+import { FabricAccessControl } from "#interaction/FabricAccessControl.js";
 import { PeerAddress } from "#peer/PeerAddress.js";
 import { Session } from "#session/Session.js";
 import { CaseAuthenticatedTag, FabricId, FabricIndex, GroupId, NodeId, VendorId } from "#types";
@@ -62,8 +62,8 @@ export class Fabric {
     readonly operationalCert: Uint8Array;
     readonly #keyPair: Key;
     readonly #sessions = new Set<Session>();
-    readonly #groupManager: FabricGroupsManager;
-    readonly #aclManager: FabricAccessControlManager;
+    readonly #groupManager: FabricGroups;
+    readonly #aclManager: FabricAccessControl;
     #label: string;
     #removeCallbacks = new Array<() => MaybePromise<void>>();
     #persistCallback: ((isUpdate?: boolean) => MaybePromise<void>) | undefined;
@@ -84,8 +84,8 @@ export class Fabric {
         this.operationalCert = config.operationalCert;
         this.#label = config.label;
         this.#keyPair = PrivateKey(config.keyPair);
-        this.#aclManager = new FabricAccessControlManager(this);
-        this.#groupManager = new FabricGroupsManager(this);
+        this.#aclManager = new FabricAccessControl(this);
+        this.#groupManager = new FabricGroups(this);
     }
 
     get config(): Fabric.Config {