+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The main work (all changes without a GitHub username in brackets in the below li
 -   @matter/node
     - Enhancement: Finalizes behavior event `interactionBegin` to fire at the begin of an interaction as soon as the datasource is about to be changed on a behavior
     - Enhancement: Finalizes behavior event `interactionEnd` to fire at the end of an interaction when all logic is done and executed
-    - Enhancement: Finalizes behavior event `datasourceChanged` to fire when an interaction changed the state of the behavior. It fires at the end after all concrete `$Changed` events are sent out
+    - Enhancement: Finalizes behavior event `stateChanged` to fire when an interaction changed the state of the behavior. It fires at the end after all concrete `$Changed` events are sent out
 
 -   @matter/protocol
     - Enhancement: Exposed reading cached ClusterClient attributes via `get*AttributeFromCache()` method